trying fix visit counter

+++ student/yuhongy_student_turtle.cpp
@@ -24,7 +24,7 @@ turtleMove studentTurtleStep(bool bumped) { return MOVE; }
 // the maze!)
 const int32_t TIMEOUT = 20; 	 // bigger number slows down simulation so you can see what's happening
 int32_t wait;   				 // w: countdown time.
-typedef double Cord;			 // Define the Coordinate type
+typedef int32_t Cord;			 // Define the Coordinate type
 struct Point2D {
     Cord x;
     Cord y;
@@ -154,7 +154,7 @@ bool studentMoveTurtle(QPointF &pos_, int &nw_or) {
 		bool bp = bumped(startPoint.x, startPoint.y, endPoint.x, endPoint.y);  // if there is a bump (boolean)
 		aend = atend(pos_.x(), pos_.y()); 									   // if arrvies at end (boolean)
 		ROS_INFO("Current state: %d, Orientation: %d", cs, nw_or);
-		ROS_INFO("Current position: x = %f, y = %f", pos_.x(), pos_.y());
+		ROS_INFO("Current position: x = %d, y = %d", pos_.x(), pos_.y());
 
 		//left hand rule
 		switch(nw_or) {
